requirements.txt
torch
torchvision
scikit-learn
kagglehub
Pillow